refactor: move ApiError, ApiResponse, and PaginationMeta to utils package and update imports
