+++ src/main/java/dev/araopj/hrplatformapi/audit/controller/AuditController.java
@@ -3,8 +3,8 @@ package dev.araopj.hrplatformapi.audit.controller;
 import dev.araopj.hrplatformapi.audit.dto.AuditDto;
 import dev.araopj.hrplatformapi.audit.model.Audit;
 import dev.araopj.hrplatformapi.audit.service.AuditService;
-import dev.araopj.hrplatformapi.dto.ApiError;
-import dev.araopj.hrplatformapi.dto.ApiResponse;
+import dev.araopj.hrplatformapi.utils.ApiError;
+import dev.araopj.hrplatformapi.utils.ApiResponse;
 import dev.araopj.hrplatformapi.utils.Mapper;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
