+++ src/main/java/dev/araopj/hrplatformapi/utils/ApiError.java
@@ -1,4 +1,4 @@
-package dev.araopj.hrplatformapi.dto;
+package dev.araopj.hrplatformapi.utils;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import jakarta.validation.constraints.NotNull;
